+++ e2e/packages/sync-test/data/numberListSystem.ts
@@ -6,9 +6,9 @@ export async function push(page: Page, num: number) {
 }
 
 export async function pushRange(page: Page, start: number, end: number) {
-  await callWorld(page, "pushRange", [start, end, { gasLimit: 20_000_000 }]);
+  await callWorld(page, "pushRange", [start, end]);
 }
 
 export async function pop(page: Page) {
-  await callWorld(page, "pop", []);
+  await callWorld(page, "pop");
 }