+++ packages/common/tsup.config.ts
@@ -3,12 +3,13 @@ import { defineConfig } from "tsup";
 export default defineConfig({
   entry: {
     index: "src/index.ts",
+    chains: "src/chains/index.ts",
     codegen: "src/codegen/index.ts",
+    deprecated: "src/deprecated/index.ts",
+    errors: "src/errors/index.ts",
     foundry: "src/foundry/index.ts",
     "type-utils": "src/type-utils/index.ts",
     utils: "src/utils/index.ts",
-    errors: "src/errors/index.ts",
-    chains: "src/chains/index.ts",
   },
   target: "esnext",
   format: ["esm"],