+++ packages/store-sync/src/recs/defineInternalComponents.ts
@@ -1,4 +1,4 @@
-import { World, defineComponent, Type } from "@latticexyz/recs";
+import { World, defineComponent, Type, Component, Schema } from "@latticexyz/recs";
 import { Table } from "../common";
 import { StoreComponentMetadata } from "./common";
 
@@ -8,7 +8,14 @@ export function defineInternalComponents(world: World) {
     TableMetadata: defineComponent<{ table: Type.T }, StoreComponentMetadata, Table>(
       world,
       { table: Type.T },
-      { metadata: { keySchema: {}, valueSchema: {} } }
+      {
+        metadata: {
+          componentName: "TableMetadata",
+          tableName: "recs:TableMetadata",
+          keySchema: {},
+          valueSchema: {},
+        },
+      }
     ),
     SyncProgress: defineComponent(
       world,
@@ -16,10 +23,17 @@ export function defineInternalComponents(world: World) {
         step: Type.String,
         message: Type.String,
         percentage: Type.Number,
+        latestBlockNumber: Type.BigInt,
+        lastBlockNumberProcessed: Type.BigInt,
       },
       {
-        metadata: { keySchema: {}, valueSchema: {} },
+        metadata: {
+          componentName: "SyncProgress",
+          tableName: "recs:SyncProgress",
+          keySchema: {},
+          valueSchema: {},
+        },
       }
     ),
-  };
+  } as const satisfies Record<string, Component<Schema, StoreComponentMetadata>>;
 }