+++ packages/store-sync/src/recs/index.ts
@@ -3,5 +3,7 @@ export * from "./decodeEntity";
 export * from "./encodeEntity";
 export * from "./entityToHexKeyTuple";
 export * from "./hexKeyTupleToEntity";
+export * from "./isStoreComponent";
 export * from "./recsStorage";
+export * from "./singletonEntity";
 export * from "./syncToRecs";